Two Room Library Examples.
Two ways to use the RoomLibrary and RoomLibrary_HashMap.
Give some feedback on what could be better to use. In the hashmap you're
able to give a keyvalue to find any room object. parsing the arraylist
used in RoomLibrary may be harder to do to find a specific room object.