Backpack, Backpack!!
We now have a working inventory class. Items are added through its
method. to a hashmap. this wont include the getters just yet for each
individual one. This is just a simple example of its use. I create three
items in the controller. And add each to the backpack. i then ask the
backpack class to print its contents (each attribute for each item) to
the user in the console. Also if you haven't checked out the previous
roomlibrary examples they will be shown as well. above the backpack
calls.